requirements.txt
langchain 
chromadb 
fastapi 
streamlit
google-generativeai 
langchain
sentence-transformers
uvicorn